refactor: renames based on PR review
