+++ sprout/templates/projects_id_metadata/step-name-and-description.html
