+++ sprout/views/__init__.py
@@ -1,6 +1,6 @@
 """Module with all views."""
 
-from .projects_id_metadata_id.create import projects_id_metadata_create
+from .projects_id_metadata_create import projects_id_metadata_create
 from .projects_id_metadata_id_data_update import projects_id_metadata_id_data_update
 from .projects_id_metadata_view import projects_id_metadata_view
 from .projects_id_view import projects_id_view
